LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/json-module/invalid-content-type.any-expected.txt

CONSOLE MESSAGE: SyntaxError: Unexpected token ','. import call expects exactly one argument.

Harness Error (FAIL), message = SyntaxError: Unexpected token ','. import call expects exactly one argument.
